video upload, user profile, video-user connection done

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/user";
 import Video from "../models/video";
 
 export const home = async (req, res) => {
@@ -13,7 +14,9 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
     // const id = req.params.id;
     const { id } = req.params;
-    const video = await Video.findById(id);
+    // populate 는 Video 모델에 연결된 Object를 동시에 불러와준다.
+    // mongoose가 Video 안의 owner 가 User Object의 id 인 것을 알고 해당 Object 를 owner에 불러와준다.
+    const video = await Video.findById(id).populate("owner");
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
@@ -26,6 +29,10 @@ export const getEdit = async (req, res) => {
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
+    // video 의 owner 와 현재 로그인한 user 를 비교하여 수정할 수 있는 권한을 확인해줌
+    if (String(video.owner) !== String(req.session.user._id)) {
+        return res.status(403).redirect("/");
+    }
     return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
@@ -35,6 +42,9 @@ export const postEdit = async (req, res) => {
     const video = await Video.exists({_id: id});
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(req.session.user._id)) {
+        return res.status(403).redirect("/");
     }
     // video.title = title;
     // video.description = description;
@@ -52,15 +62,32 @@ export const getUpload = (req, res) => {
     return res.render("upload", { pageTitle: `Upload Video` });
 };
 export const postUpload = async (req, res) => {
-    const { title, description, hashtags } = req.body;
+    // const file = req.file;
+    // const user = req.session;
+    // const { title, description, hashtags } = req.body;
+    const {
+        session: {
+            user: { _id },
+        },
+        body: {
+            title, description, hashtags,
+        },
+        file,
+    } = req;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title: title,
             description: description,
+            fileUrl: file.path,
+            owner: _id,
             // createdAt: Date.now(),
             // default date 지정을 통해 삭제 가능
             hashtags: Video.formatHashtags(hashtags),
         });
+        // User 의 Videos Array 에 새로 등록하는 video 의 id 를 추가해줌
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch (error) {
         return res.status(400).render("upload", {
@@ -72,6 +99,13 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(req.session.user._id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
